refactor : chatRoom chatMessage 위치 변경
