--- fitConnect/src/main/java/com/example/fitconnect/repository/chat/ChatRoomRepository.java
@@ -1,8 +1,0 @@
-package com.example.fitconnect.repository.chat;
-
-import com.example.fitconnect.domain.chat.domain.ChatRoom;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ChatRoomRepository extends JpaRepository<ChatRoom,Long> {
-
-}
