+++ fitConnect/src/main/java/com/example/fitconnect/repository/chat/chatMessage/ChatMessageRepository.java
@@ -1,4 +1,4 @@
-package com.example.fitconnect.repository.chat;
+package com.example.fitconnect.repository.chat.chatMessage;
 
 import com.example.fitconnect.domain.chat.domain.ChatMessage;
 import org.springframework.data.jpa.repository.JpaRepository;
