+++ fitConnect/src/main/java/com/example/fitconnect/service/chat/chatMessage/ChatMessageDeleteService.java
@@ -4,7 +4,7 @@ package com.example.fitconnect.service.chat.chatMessage;
 import com.example.fitconnect.config.error.ErrorMessages;
 import com.example.fitconnect.config.exception.EntityNotFoundException;
 import com.example.fitconnect.domain.chat.domain.ChatMessage;
-import com.example.fitconnect.repository.chat.ChatMessageRepository;
+import com.example.fitconnect.repository.chat.chatMessage.ChatMessageRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
