+++ fitConnect/src/test/java/com/example/fitconnect/controller/chat/ChatRoomControllerTest.java
@@ -7,7 +7,7 @@ import com.example.fitconnect.domain.chat.dto.ChatRoomRegistrationDto;
 import com.example.fitconnect.domain.chat.dto.ChatRoomUpdateDto;
 import com.example.fitconnect.domain.event.domain.ExerciseEvent;
 import com.example.fitconnect.domain.user.domain.User;
-import com.example.fitconnect.repository.chat.ChatRoomRepository;
+import com.example.fitconnect.repository.chat.chatRoom.ChatRoomRepository;
 import com.example.fitconnect.service.chat.chatRoom.ChatRoomCreationService;
 import com.example.fitconnect.service.chat.chatRoom.ChatRoomDeleteService;
 import com.example.fitconnect.service.chat.chatRoom.ChatRoomUpdateService;
